Thank you so much for your interest in Ninja Theory for the position of Mid / Senior Gameplay Programmer.
I’m pleased to let you know that you have been selected for the next stage in the recruitment process which is a test.

Please find the test attached to this email. We kindly ask that you complete this as part of your application and return this to us by 9am Wednesday 23rd March.

When you are finished, please zip up all your code. You should not include build intermediates or executables in the zip file - we will build your code when we receive your completed test. Please submit your test by replying to this email.

If you will not have availability to complete the assignment in the given time frame or you have any further questions, please do not hesitate to contact me.

Best wishes,

Beth Cates